# code/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parking foundation — local development",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/facilities/{facility_id}/relationships": {
      "get": {
        "summary": "Relationships",
        "operationId": "relationships_api_v1_facilities__facility_id__relationships_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/customers": {
      "post": {
        "summary": "Create Customer",
        "operationId": "create_customer_api_v1_facilities__facility_id__relationships_customers_post",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/customers/{user_id}": {
      "patch": {
        "summary": "Change Customer",
        "operationId": "change_customer_api_v1_facilities__facility_id__relationships_customers__user_id__patch",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/vehicles": {
      "post": {
        "summary": "Create Vehicle",
        "operationId": "create_vehicle_api_v1_facilities__facility_id__relationships_vehicles_post",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/vehicles/{vehicle_id}": {
      "patch": {
        "summary": "Change Vehicle",
        "operationId": "change_vehicle_api_v1_facilities__facility_id__relationships_vehicles__vehicle_id__patch",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/vehicles/{vehicle_id}/customer": {
      "put": {
        "summary": "Change Vehicle User",
        "operationId": "change_vehicle_user_api_v1_facilities__facility_id__relationships_vehicles__vehicle_id__customer_put",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUserChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/vehicle-objects/{object_id}": {
      "put": {
        "summary": "Change Vehicle Object",
        "operationId": "change_vehicle_object_api_v1_facilities__facility_id__relationships_vehicle_objects__object_id__put",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectMappingChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/relationships/person-objects/{object_id}": {
      "put": {
        "summary": "Change Person Object",
        "operationId": "change_person_object_api_v1_facilities__facility_id__relationships_person_objects__object_id__put",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonMappingChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerChange": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          },
          "display_alias": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Alias"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "CustomerChange"
      },
      "CustomerCreate": {
        "properties": {
          "display_alias": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Display Alias"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_alias",
          "reason"
        ],
        "title": "CustomerCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObjectMappingChange": {
        "properties": {
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Run Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "registered_vehicle_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Registered Vehicle Id"
          },
          "mapping_status": {
            "type": "string",
            "enum": [
              "verified",
              "uncertain",
              "unmapped"
            ],
            "title": "Mapping Status"
          },
          "mapping_source": {
            "type": "string",
            "enum": [
              "demo_config",
              "reviewed"
            ],
            "title": "Mapping Source"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "expected_version",
          "mapping_status",
          "mapping_source",
          "reason"
        ],
        "title": "ObjectMappingChange"
      },
...
      "PersonMappingChange": {
        "properties": {
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Run Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposed",
              "uncertain",
              "verified",
              "unmapped"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "enum": [
              "demo_config",
              "reviewed"
            ],
            "title": "Source"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "expected_version",
          "status",
          "source",
          "reason"
        ],
        "title": "PersonMappingChange"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleChange": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          },
          "display_alias": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Alias"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "VehicleChange"
      },
      "VehicleCreate": {
        "properties": {
          "display_alias": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Display Alias"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_alias",
          "reason"
        ],
        "title": "VehicleCreate"
      },
...
      "VehicleUserChange": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "VehicleUserChange"
      }
    }
  }
}